feat(discover) Add deprecation warning for events and discover v1
These views will be removed in the near future for most accounts as we
transition over to the new am plans and feature sets. The plans in
production that have access to events and discover also have access to
discover-basic and or discover-query

There is no firm date for removal so that we trap future us.
