+++ src/sentry/static/sentry/app/views/events/index.jsx
@@ -6,16 +6,21 @@ import isEqual from 'lodash/isEqual';
 import {loadOrganizationTags} from 'app/actionCreators/tags';
 import {getParams} from 'app/components/organizations/globalSelectionHeader/getParams';
 import {t} from 'app/locale';
+import Alert from 'app/components/alert';
 import FeatureBadge from 'app/components/featureBadge';
 import Feature from 'app/components/acl/feature';
+import Link from 'app/components/links/link';
 import GlobalSelectionHeader from 'app/components/organizations/globalSelectionHeader';
 import LightWeightNoProjectMessage from 'app/components/lightWeightNoProjectMessage';
+import {IconWarning} from 'app/icons';
 import SentryTypes from 'app/sentryTypes';
 import PageHeading from 'app/components/pageHeading';
 import withApi from 'app/utils/withApi';
 import withGlobalSelection from 'app/utils/withGlobalSelection';
 import withOrganization from 'app/utils/withOrganization';
+import {getDiscoverLandingUrl} from 'app/utils/discover/urls';
 import {PageContent, PageHeader} from 'app/styles/organization';
+import space from 'app/styles/space';
 
 import SearchBar from './searchBar';
 
@@ -55,6 +60,7 @@ class EventsContainer extends React.Component {
 
   render() {
     const {organization, location, children, selection} = this.props;
+    const discoverUrl = getDiscoverLandingUrl(organization);
 
     return (
       <Feature
@@ -70,6 +76,13 @@ class EventsContainer extends React.Component {
                   <HeaderTitle>
                     {t('Events')} <FeatureBadge type="beta" />
                   </HeaderTitle>
+                </PageHeader>
+                <div>
+                  <Alert type="warning" icon={<IconWarning size="sm" />}>
+                    The events view is deprecated. Check out our new visualization and
+                    querying capabilities in{' '}
+                    <Link to={discoverUrl}>the new Discover</Link> today.
+                  </Alert>
                   <StyledSearchBar
                     organization={organization}
                     projectIds={selection.projects}
@@ -79,7 +92,7 @@ class EventsContainer extends React.Component {
                     )}
                     onSearch={this.handleSearch}
                   />
-                </PageHeader>
+                </div>
                 {children}
               </Body>
             </LightWeightNoProjectMessage>
@@ -104,4 +117,5 @@ const HeaderTitle = styled(PageHeading)`
 
 const StyledSearchBar = styled(SearchBar)`
   flex: 1;
+  margin-bottom: ${space(2)};
 `;
