+++ src/sentry/static/sentry/app/views/eventsV2/breadcrumb.tsx
@@ -5,8 +5,7 @@ import {t} from 'app/locale';
 import {Event, Organization} from 'app/types';
 import EventView from 'app/utils/discover/eventView';
 import Breadcrumbs, {Crumb} from 'app/components/breadcrumbs';
-
-import {getDiscoverLandingUrl} from './utils';
+import {getDiscoverLandingUrl} from 'app/utils/discover/urls';
 
 type DefaultProps = {
   event: Event | undefined;
