+++ tests/js/spec/views/eventsV2/utils.spec.jsx
@@ -5,7 +5,6 @@ import {
   decodeColumnOrder,
   pushEventViewToLocation,
   getExpandedResults,
-  getDiscoverLandingUrl,
   downloadAsCsv,
 } from 'app/views/eventsV2/utils';
 import {COL_WIDTH_UNDEFINED} from 'app/components/gridEditable';
@@ -435,18 +434,6 @@ describe('getExpandedResults()', function() {
     const event = {'project.name': 'whoosh'};
     const result = getExpandedResults(view, {}, event);
     expect(result.query).toEqual('project.name:whoosh');
-  });
-});
-
-describe('getDiscoverLandingUrl', function() {
-  it('is correct for with discover-query and discover-basic features', function() {
-    const org = TestStubs.Organization({features: ['discover-query', 'discover-basic']});
-    expect(getDiscoverLandingUrl(org)).toBe('/organizations/org-slug/discover/queries/');
-  });
-
-  it('is correct for with only discover-basic feature', function() {
-    const org = TestStubs.Organization({features: ['discover-basic']});
-    expect(getDiscoverLandingUrl(org)).toBe('/organizations/org-slug/discover/results/');
   });
 });
